refactor: consolidate Gist setup and sync
