+++ services/document-gist-sync.ts
@@ -1,4 +1,6 @@
 import { decideGistSync, type GistSyncStatus, type SyncResult } from '@/domain/gist-sync';
+import { translations } from '@/i18n';
+import { detectBrowserLanguage } from '@/infrastructure/browser/language';
 import { createGitHubClient, GIST_FILENAME } from '@/infrastructure/github/client';
 import { readGistConnection } from '@/infrastructure/storage/gist-connection';
 import {
@@ -7,6 +9,7 @@ import {
   replaceLearningDocument,
 } from '@/infrastructure/storage/learning-document';
 import { readSyncMetadata, writeSyncStatus } from '@/infrastructure/storage/sync-metadata';
+import { createGistFromBackup } from './gist-sync';
 
 // Prepared for the coordinated runtime activation in #378.
 // Keep the legacy workflow authoritative until then; #379 removes its duplicate
@@ -93,4 +96,26 @@ export async function triggerGistSync(): Promise<SyncResult> {
   } finally {
     syncInProgress = false;
   }
+}
+
+export async function createNewGist(): Promise<{ gistId: string }> {
+  const config = await readGistConnection();
+  if (!config.pat) {
+    throw new Error('PAT is required to create a gist');
+  }
+
+  const github = createGitHubClient(config.pat);
+  const document = await readLearningDocument();
+  if (!document) {
+    throw new Error('Learning document is not initialized');
+  }
+
+  // Resolve from the exported snapshot; the translation storage adapter would
+  // read a second document that could have a different language.
+  const language = document.settings.language ?? detectBrowserLanguage();
+  return createGistFromBackup(
+    github,
+    JSON.stringify(document, null, 2),
+    translations[language].settings.gistSync.gistDescription
+  );
 }
